% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project.R
\name{beta_hat}
\alias{beta_hat}
\title{Optimize function}
\usage{
beta_hat(x, y)
}
\arguments{
\item{x}{a matrix of numerical coefficients of beta}

\item{y}{a numerical vector}
}
\description{
Optimize function
}
