% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project.R
\name{bhat.predict}
\alias{bhat.predict}
\title{Prediction function}
\usage{
bhat.predict(model, x)
}
\arguments{
\item{x}{a matrix of numerical coefficients of beta}
}
\value{
beta hat
}
\description{
Prediction function
}
