% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project.R
\name{Boostrap}
\alias{Boostrap}
\title{bootstrap function
boostrap confidence interval}
\usage{
Boostrap(x, y, alpha = 0.05, R = 20, seed = NULL, ...)
}
\arguments{
\item{x}{a matrix of numerical coefficients of beta}

\item{y}{a numerical vector}

\item{seed}{accounts for randomness and will allow for the same generated numbers}

\item{...}{}
}
\description{
bootstrap function
boostrap confidence interval
}
