% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project.R
\name{grad}
\alias{grad}
\title{Gradient descent function
Constraint function}
\usage{
grad(beta, x, y)
}
\arguments{
\item{beta}{the coefficient vector}

\item{x}{a matrix of numerical coefficients of beta}

\item{y}{a numerical vector}
}
\description{
Gradient descent function
Constraint function
}
