% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project.R
\name{model_predict}
\alias{model_predict}
\title{Title}
\usage{
model_predict(Z, x, y)
}
\arguments{
\item{x}{a matrix of numerical coefficients of beta}

\item{y}{a numerical vector}
}
\description{
Title
}
