% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project.R
\name{obj_fn}
\alias{obj_fn}
\title{Objective function
Finds the beta hat that will minimize the objective function}
\usage{
obj_fn(beta, x, y)
}
\arguments{
\item{beta}{the coefficient vector}

\item{x}{a matrix of numerical coefficients of beta}

\item{y}{a numerical vector}
}
\description{
Objective function
Finds the beta hat that will minimize the objective function
}
