% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_Project.R
\name{Pi}
\alias{Pi}
\title{Pi function
Define the probability of success or failure}
\usage{
Pi(m)
}
\arguments{
\item{m}{numerical data set to be fitted to a logistic regression curve}
}
\value{
logistic regression curve
}
\description{
Pi function
Define the probability of success or failure
}
\examples{
Pi(data)
}
